Rename hackathon package to cfmanager
