+++ cf-psb/psb-service/src/main/java/com/emc/ocopea/cfpsb/CfPsbSingleton.java
@@ -7,9 +7,9 @@ import com.emc.microservice.singleton.ServiceLifecycle;
 import com.emc.microservice.webclient.WebAPIResolver;
 import com.emc.microservice.webclient.WebApiResolverBuilder;
 import com.emc.ocopea.devtools.checkstyle.NoJavadoc;
-import com.emc.ocopea.hackathon.CFConnection;
-import com.emc.ocopea.hackathon.CloudFoundryClientManagedResource;
-import com.emc.ocopea.hackathon.CloudFoundryClientResourceDescriptor;
+import com.emc.ocopea.cfmanager.CFConnection;
+import com.emc.ocopea.cfmanager.CloudFoundryClientManagedResource;
+import com.emc.ocopea.cfmanager.CloudFoundryClientResourceDescriptor;
 import com.emc.ocopea.psb.DeployAppServiceManifestDTO;
 import com.emc.ocopea.psb.PSBAppServiceInstanceDTO;
 import com.emc.ocopea.psb.PSBAppServiceStatusEnumDTO;
