+++ cf-tools/cf-external-resource-manager/src/main/java/com/emc/ocopea/cfmanager/CFConnection.java
@@ -1,6 +1,6 @@
 // Copyright (c) [2017] Dell Inc. or its subsidiaries. All Rights Reserved.
  
-package com.emc.ocopea.hackathon;
+package com.emc.ocopea.cfmanager;
 
 import org.cloudfoundry.client.CloudFoundryClient;
 import org.cloudfoundry.operations.CloudFoundryOperations;
