+++ cf-tools/cf-external-resource-manager/src/main/java/com/emc/ocopea/cfmanager/CloudFoundryClientResourceDescriptor.java
@@ -1,6 +1,6 @@
 // Copyright (c) [2017] Dell Inc. or its subsidiaries. All Rights Reserved.
  
-package com.emc.ocopea.hackathon;
+package com.emc.ocopea.cfmanager;
 
 import com.emc.microservice.resource.ResourceDescriptor;
