+++ cf-tools/cf-external-resource-manager/src/main/resources/META-INF/services/com.emc.microservice.resource.ExternalResourceManager
@@ -1,1 +1,1 @@
-com.emc.ocopea.hackathon.CloudFoundryClientResourceManager
+com.emc.ocopea.cfmanager.CloudFoundryClientResourceManager
